refactor: homogenize record 006 schema
by moving the 00-position spec into the individual type definitions.

If there was a reason to split it out like it had been, it's not obvious.
It's easier to parse this field now, without the need to specially
handle edge cases.

The concept of an "All Materials" type appears to be a (code-side)
implementation detail; I see no mention of it as an independent concept
in the MARC21 specification.