feat: allow search to customize medium and format
I originally thought search was improperly escaping terms,
but it was actually performing a legitimately empty search;
the problem was by having medium hard-coded to blu-ray, searches for
items not explicitly tagged came back empty.

We no longer hard-code the medium to blu-ray;
the default value is now None, but this can be overridden.

We still assume searches are for video content,
but this can now be overridden too.